Добавляет маршрутизацию для всех внутренних маршрутов в профиле пользователя
