+++ src/App.tsx
@@ -2,8 +2,6 @@ import { Route, Routes } from 'react-router-dom'
 import Auth0ProviderWithNavigate from './user/auth0/auth0-provider'
 import Login from './user/components/Login/Login'
 import Register from './user/components/Register/Register'
-import Profile from './user/components/Profile/Profile'
-import { ProfileCourses } from './user/components/ProfileCourses/ProfileCourses.tsx'
 import AuthCallback from './user/components/AuthCallback'
 import ProtectedRoute from './user/components/ProtectedRoute'
 import PasswordReset from './user/components/PasswordReset/PasswordReset'
@@ -18,6 +16,13 @@ import { ThemeProvider } from '@mui/material/styles'
 import CssBaseline from '@mui/material/CssBaseline'
 import { authTheme } from './user/stylesObj'
 import AdminPanel from './pages/AdminPanel'
+import { ProfileLayout } from './user/components/ProfileLayout/ProfileLayout.tsx'
+import { ProfileCoursesPage } from './user/components/ProfilePages/ProfileCoursesPage.tsx'
+import { ProfilePage } from './user/components/ProfilePages/ProfilePage.tsx'
+import { ProfileExercisesPage } from './user/components/ProfilePages/ProfileExercisesPage.tsx'
+import { ProfileGrammarPage } from './user/components/ProfilePages/ProfileGrammarPage.tsx'
+import { ProfileGamesPage } from './user/components/ProfilePages/ProfileGamesPage.tsx'
+import { ProfileSettingsPage } from './user/components/ProfilePages/ProfileSettingsPage.tsx'
 
 const App: React.FC = () => {
   return (
@@ -32,18 +37,17 @@ const App: React.FC = () => {
               path="/profile"
               element={
                 <ProtectedRoute>
-                  <Profile />
+                  <ProfileLayout />
                 </ProtectedRoute>
               }
-            />
-            <Route
-              path="/profile-courses"
-              element={
-                <ProtectedRoute>
-                  <ProfileCourses />
-                </ProtectedRoute>
-              }
-            />
+            >
+              <Route index element={<ProfilePage />} />
+              <Route path="courses" element={<ProfileCoursesPage />} />
+              <Route path="exercises" element={<ProfileExercisesPage />} />
+              <Route path="grammar" element={<ProfileGrammarPage />} />
+              <Route path="games" element={<ProfileGamesPage />} />
+              <Route path="settings" element={<ProfileSettingsPage />} />
+            </Route>
             <Route path="/reset-password" element={<PasswordReset />} />
             <Route path="/auth-callback" element={<AuthCallback />} />
             <Route path="/courses" element={<Courses />} />
