+++ src/user/components/Profile/CourseCard.tsx
