+++ src/user/components/Profile/mockDataSlider.ts
