--- src/user/components/ProfileCourses/constants.ts
@@ -1,66 +1,0 @@
-export const mockDataLanguageCourses = [
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Английский',
-    level: 'A1',
-    description: 'Идеально для начинающих',
-    amountTime: 20,
-    progress: null,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Французский',
-    level: 'A1',
-    description: 'Расширьте знания',
-    amountTime: 20,
-    progress: 55,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Испанский',
-    level: 'A1',
-    description: 'Идеально для начинающих',
-    amountTime: 20,
-    progress: 77,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Португальский',
-    level: 'A1',
-    description: 'Расширьте знания',
-    amountTime: 20,
-    progress: 44,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Немецкий',
-    level: 'A1',
-    description: 'Идеально для начинающих',
-    amountTime: 20,
-    progress: null,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Китайский',
-    level: 'A1',
-    description: 'Расширьте знания',
-    amountTime: 20,
-    progress: 11,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Арабский',
-    level: 'A1',
-    description: 'Идеально для начинающих',
-    amountTime: 20,
-    progress: null,
-  },
-  {
-    id: window.crypto.randomUUID(),
-    language: 'Русский',
-    level: 'A1',
-    description: 'Расширьте знания',
-    amountTime: 20,
-    progress: 88,
-  },
-]
