+++ src/user/components/ProfilePages/ProfileExercisesPage.tsx
@@ -1,0 +1,17 @@
+import { Box } from '@mui/material'
+import { Typography } from '@mui/material'
+
+const ProfileExercisesPage = () => {
+  return (
+    <Box sx={{ padding: 3 }}>
+      <Typography variant="h4" gutterBottom>
+        Упражнения
+      </Typography>
+      <Typography variant="body1">
+        Страница с упражнениями. Содержимое будет добавлено позже.
+      </Typography>
+    </Box>
+  )
+}
+
+export { ProfileExercisesPage }
