+++ src/user/components/ProfilePages/ProfileGamesPage.tsx
@@ -1,0 +1,17 @@
+import { Box } from '@mui/material'
+import { Typography } from '@mui/material'
+
+const ProfileGamesPage = () => {
+  return (
+    <Box sx={{ padding: 3 }}>
+      <Typography variant="h4" gutterBottom>
+        Игры
+      </Typography>
+      <Typography variant="body1">
+        Страница с обучающими играми. Содержимое будет добавлено позже.
+      </Typography>
+    </Box>
+  )
+}
+
+export { ProfileGamesPage }
