+++ src/user/components/ProfilePages/ProfileGrammarPage.tsx
@@ -1,0 +1,17 @@
+import { Box } from '@mui/material'
+import { Typography } from '@mui/material'
+
+const ProfileGrammarPage = () => {
+  return (
+    <Box sx={{ padding: 3 }}>
+      <Typography variant="h4" gutterBottom>
+        Грамматика
+      </Typography>
+      <Typography variant="body1">
+        Страница с материалами по грамматике. Содержимое будет добавлено позже.
+      </Typography>
+    </Box>
+  )
+}
+
+export { ProfileGrammarPage }
