+++ src/user/components/ProfilePages/ProfileSettingsPage.tsx
@@ -1,0 +1,17 @@
+import { Box } from '@mui/material'
+import { Typography } from '@mui/material'
+
+const ProfileSettingsPage = () => {
+  return (
+    <Box sx={{ padding: 3 }}>
+      <Typography variant="h4" gutterBottom>
+        Настройки
+      </Typography>
+      <Typography variant="body1">
+        Страница настроек профиля. Содержимое будет добавлено позже.
+      </Typography>
+    </Box>
+  )
+}
+
+export { ProfileSettingsPage }
